Use a regex to debounce messages. (#3)
* Use a regex to debounce messages.

This makes debounce slightly safer as it specifically checks the END of
the message text, to avoid scenarios where a message may contain '(x'
(our indication of a duplicate message) in the message body but NOT at
the end. E.g. - "Host (x1234abc) is on fire".

* Compile debounce regex at module level.

Don't compile the regex on every `debounce` call - instead do this at
the module level. Update the pattern to use a named group to pull the
last count.